Move JtsShape to shared
