+++ modules/math/jvm/src/main/scala/gsp/math/geom/jts/jvm/syntax/package.scala
@@ -1,14 +1,10 @@
 // Copyright (c) 2016-2020 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package gsp.math.geom.jts
+package gsp.math.geom.jts.jvm
 
 package object syntax {
 
-  // Syntax used in the JTS implementation only.
-
-  object all extends ToAngleOps
-                with ToOffsetOps
+  object all extends ToAwtOps
 
 }
-
