+++ modules/tests/jvm/src/main/scala/gsp/math/geom/jts/demo/JtsDemo.scala
@@ -6,6 +6,7 @@ package demo
 
 import gsp.math.Angle
 import gsp.math.geom.ShapeExpression
+import gsp.math.geom.jts.jvm.syntax.awt._
 import gsp.math.geom.jts.interpreter._
 import gsp.math.geom.syntax.shapeexpression._
 import gsp.math.syntax.int._
